refactor: new config name
