+++ internal/otherconfig/getters.go
@@ -1,4 +1,4 @@
-package configuration
+package otherconfig
 
 import "github.com/knadh/koanf/parsers/json"
